pyproject.toml
[tool.poetry]
name = "epicteller"
version = "0.1.0"
description = "TRPG Web & IM Toolkits"
authors = ["Nitori <nitori@ikazuchi.cn>"]
license = "GPL3"

[tool.poetry.dependencies]
python = "^3.9"
click = ">=5.0"
SQLAlchemy = "<1.4"
supervisor = ">=4.0.3"
environs = ">=4.2.0"
lark-parser = "*"
epicteller-datum = ">=0.4.1"
ipython = ">=7.7.0"
aiomysql = "*"
aioredis = "*"
pybase62 = "*"
aiokafka = "*"
bcrypt = "*"
httpx = "*"
cos-python-sdk-v5 = "*"
pillow = ">=8.1.1"
fastapi = "*"
uvicorn = {extras = ["standard"], version = "*"}
python-jose = "*"
dataclasses-json = "*"
email-validator = "*"
nonebot2 = "==2.0.0a10"
lru-dict = "*"
loguru = "*"
pexpect = "*"
sentry-sdk = "*"
nonebot-plugin-sentry = "*"
gitpython = "*"
cryptography = "*"
tldextract = "*"

[tool.poetry.dev-dependencies]
pytest = ">=4.6.3"
WebTest = ">=2.0.33"
factory-boy = ">=2.12.*"
pdbpp = ">=0.10.0"
flake8 = ">=3.7.7"
flake8-blind-except = ">=0.1.1"
flake8-debugger = ">=3.1.0"
flake8-docstrings = ">=1.3.0"
flake8-isort = ">=2.7.0"
pep8-naming = ">=0.8.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
